pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-app"
version = "0.1.0"
description = "A Python CLI that converts text to speech using OpenAI's TTS API and merges audio files."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "tts-app contributors"},
]
keywords = ["tts", "text-to-speech", "openai", "audio", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/tts-app"
Repository = "https://github.com/YOUR_USERNAME/tts-app"
Issues = "https://github.com/YOUR_USERNAME/tts-app/issues"

[project.scripts]
tts = "tts:main"

[tool.setuptools]
py-modules = ["tts"]
